Update queries.c - Workaround for LW8's "*?Resolution" query
Technically comments in PPD files starting with "*?" are actual bits of Postscript code a driver can send to a printer. The PPD parsing code in papd, normally ignores these. Printer drivers normally don't send queries like this to a printer or spooler, expect for LaserWriter 8. So we make an exception here and change the '?' to a '*'.