[papd] Allow papd to return a printer resolution with LaserWriter 8 driver. (#37)
* Update ppd.c - Add Resolution as a supported query

* Update queries.c - Workaround for LW8's "*?Resolution" query

Technically comments in PPD files starting with "*?" are actual bits of Postscript code a driver can send to a printer. The PPD parsing code in papd, normally ignores these. Printer drivers normally don't send queries like this to a printer or spooler, expect for LaserWriter 8. So we make an exception here and change the '?' to a '*'.

* Update print_cups.c

Remove the '?' from the Resolution line written to the PPD file. We can eventually use IPP queries to get the real printer device resolution from the printer, but we'll leave it set to a sensible default for now.

* Update print_cups.c

This is cosmetic. We don't support PostScript Fax.